+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.2.0-rc.2
 ## 03/xx/2017
 
+1. [](#improved)
+    * Updated vendor libraries to latest
 1. [](#bugfix)
     * Fixed an issue with theme inheritance and hyphenated base themes [#1353](https://github.com/getgrav/grav/issues/1353)
     * Fixed an issue when trying to use an `@2x` derivative on a non-image media file [#1341](https://github.com/getgrav/grav/issues/1341)